Assignment 3, Task 8
Your Set game should also report (mis)matches like Required Task #5 in
the last assignment, but you’ll have to enhance this feature (to use
NSAttributedString) to make it capable of working for both the Set and
Playing Card games.

+++ Matchismo/Model/CardMatchingGame.h
@@ -21,6 +21,6 @@
 @property (nonatomic, readonly) NSInteger score;
 @property (nonatomic) NSUInteger maxMatchingCards;
 @property (nonatomic, readonly) NSArray *lastChosenCards;
-@property (nonatomic, readonly) NSInteger lastScore;
+@property (nonatomic, readonly) long lastScore;
 
 @end
